Add tree node collapse/expand notifications to metals.
Revert behaviour to of return key in treeview to execute action, tab expands.
Classes can be expanded as well as a command executed, so we must be explicit
about expand and execute actions.

Add behaviour to send node expanded/collapsed to metals based on treemacs
extension added using after-expand, after-collapse.

refactor treeview-get-children and rename function to make it explicit
we're dealing with current-node.

Add debug helper to query properties of current node in tree and write to
mini buffer.