Treeview node update fixes.
Attempt to fix spurious errors during node update which could
be within treemacs - do-update-node (1+ nil) (depth?).

When finding a node if we fail to find the node in the cache,
return nil but remove the node from the cache - it's no longer
in the treemacs tree.

lsp--metals-treeview-get-children only cache nodes if we have
children and a valid current node.

lsp--metals-treeview-get-children-current-node, make sure we
have a valid current code before attempting to retrieve children.