Support ability to hide/show treeview.
Treeview buffer names now start with a space so they are hidden
from the buffer list.
Add features to show/hide treeview rather than delete it - this
retains state of the tree when hidden.
These features are a precursor to supporting the ability to
switch treeview when switching between files in different
workspaces.