Switch between active treeviews when switching workspace files.
If the user is displaying a treeview for an active workspace
and switches to a file in another workspace the treeview will
automatically change to the treeview for the new workspace.
This feature is still under test and not activated - need
to find a way of configuring an activating once per lsp-mode
session for all Metals workspaces.